refactor : 내용 변경 및, 새로운 API 추가
기존의 API로는 전체 게시물 조회가 불가 ( 특정 유저의 게시물만 조회 가능 ) 따라서 이를 분리하여 두 가지 API로 나눔
